# backend/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Sol backend api.",
        "title": "Sol API",
        "termsOfService": "https://uuree.fibo.cloud/",
        "contact": {
            "name": "API Support",
            "url": "https://www.facebook.com/uurtsaikh.n",
            "email": "vvrtsaix.n@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/auth/info": {
            "get": {
                "security": [
                    {
                        "Admin": []
                    }
                ],
                "tags": [
                    "[Admin] Auth"
                ],
                "summary": "Info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/db.Admin"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/auth/init": {
            "get": {
                "tags": [
                    "[Admin] Auth"
                ],
                "summary": "Init",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/auth/login": {
            "post": {
                "tags": [
                    "[Admin] Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/admin.LoginResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/customer/auth/info": {
            "get": {
                "security": [
                    {
                        "Customer": []
                    }
                ],
                "tags": [
                    "[Customer] Auth"
                ],
                "summary": "Info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/db.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/customer/auth/login": {
            "post": {
                "tags": [
                    "[Customer] Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/customer.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/customer.LoginResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/structs.ResponseBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.LoginInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "admin.LoginResult": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/db.Admin"
                }
            }
        },
...
        "customer.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "customer.LoginResult": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/db.User"
                }
            }
        },
...
        "db.Account": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/db.Currency"
                },
                "currency_id": {
                    "type": "integer"
                },
                "enable": {
                    "type": "boolean"
                },
                "expense_balance": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "income_balance": {
                    "type": "number"
                },
                "initial_balance": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/db.User"
                },
                "owner_id": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Transaction"
                    }
                },
                "type": {
                    "$ref": "#/definitions/db.AccountType"
                },
                "type_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/db.Workspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "db.AccountType": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Account"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meta": {
                    "$ref": "#/definitions/db.Meta"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/db.Workspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "db.Admin": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "db.Category": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_fee": {
                    "type": "boolean"
                },
                "meta": {
                    "$ref": "#/definitions/db.Meta"
                },
                "name": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Transaction"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/db.Workspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "db.Currency": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Account"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meta": {
                    "$ref": "#/definitions/db.Meta"
                },
                "name": {
                    "type": "string"
                },
                "short": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/db.Workspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "db.Meta": {
            "type": "object",
            "properties": {
                "creator": {
                    "$ref": "#/definitions/db.User"
                },
                "creator_id": {
                    "type": "integer"
                },
                "updater": {
                    "$ref": "#/definitions/db.User"
                },
                "updater_id": {
                    "type": "integer"
                }
            }
        },
        "db.Place": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/db.Meta"
                },
                "name": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Transaction"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/db.Workspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "db.Transaction": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/db.Account"
                },
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/db.Category"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meta": {
                    "$ref": "#/definitions/db.Meta"
                },
                "place": {
                    "$ref": "#/definitions/db.Place"
                },
                "place_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/db.Workspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "db.User": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Account"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_date": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspaces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.WorkspaceMember"
                    }
                }
            }
        },
        "db.Workspace": {
            "type": "object",
            "properties": {
                "account_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.AccountType"
                    }
                },
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Account"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Category"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "currencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Currency"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "invite_code": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.WorkspaceMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/db.User"
                },
                "owner_id": {
                    "type": "integer"
                },
                "places": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Place"
                    }
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Transaction"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "db.WorkspaceMember": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/db.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "workspace": {
                    "$ref": "#/definitions/db.Workspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
...
        "structs.ResponseBody": {
            "type": "object",
            "properties": {
                "body": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Admin": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Customer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Workspace": {
            "type": "apiKey",
            "name": "Workspace",
            "in": "header"
        }
    }
}